# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Quantum FX Bot API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /team/earnings-overview:
    get:
      operationId: getTeamEarningsOverview
      tags: [team]
      parameters:
        - name: period
          in: query
          schema:
            type: string
            enum: [this_month, last_month, all_time]
      responses:
        "200":
          description: Team earnings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TeamEarnings"

  # Profile
...
  /support/tickets:
    get:
      operationId: listSupportTickets
      tags: [support]
      responses:
        "200":
          description: Support tickets
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SupportTicket"
    post:
      operationId: createSupportTicket
      tags: [support]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SupportTicketInput"
      responses:
        "201":
          description: Ticket created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SupportTicket"
...
  /support/chat:
    get:
      operationId: getChatMessages
      tags: [support]
      responses:
        "200":
          description: Chat messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ChatMessage"
    post:
      operationId: sendChatMessage
      tags: [support]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendChatMessageInput"
      responses:
        "201":
          description: Message sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChatMessage"
...
components:
  schemas:
...
    ChartPoint:
      type: object
      properties:
        date:
          type: string
        label:
          type: string
        profit:
          type: number
        cumulative:
          type: number
      required: [date, label, profit, cumulative]
...
    TeamEarnings:
      type: object
      properties:
        period:
          type: string
        totalEarnings:
          type: number
        chart:
          type: array
          items:
            $ref: "#/components/schemas/ChartPoint"
      required: [period, totalEarnings, chart]

    # Profile
...
    SupportTicket:
      type: object
      properties:
        id:
          type: integer
        subject:
          type: string
        message:
          type: string
        status:
          type: string
        createdAt:
          type: string
        updatedAt:
          type: string
      required: [id, subject, message, status, createdAt, updatedAt]
    SupportTicketInput:
      type: object
      properties:
        subject:
          type: string
        message:
          type: string
        category:
          type: string
      required: [subject, message, category]
...
    ChatMessage:
      type: object
      properties:
        id: { type: integer }
        sender: { type: string }
        message: { type: string }
        createdAt: { type: string }
      required: [id, sender, message, createdAt]
    SendChatMessageInput:
      type: object
      properties:
        message: { type: string }
      required: [message]
...